fix(repo): CR.15 — restore the lint gate repo-wide, migrate web to ESLint CLI
The lint gate (AGENTS.md quality gates) was broken everywhere, not just in
web: both `next lint` and the root `eslint .` crashed loading
@typescript-eslint/await-thenable — recommended-type-checked was enabled
with no parserOptions providing type information.

Config:
- Root .eslintrc.json: parserOptions.projectService, a
  disable-type-checked override for *.test.ts / *.config.ts (files outside
  tsconfig projects), varsIgnorePattern ^_, apps/web/ excluded (lints
  itself).
- Web: migrate off deprecated `next lint` (removed in Next 16) to a flat
  eslint.config.mjs (ESLint 9) — FlatCompat next/core-web-vitals +
  typescript-eslint typed rules scoped to ts/tsx via projectService;
  legacy .eslintrc.json removed; lint script is `eslint .`.
- Root `pnpm lint` now chains the web lint — ci-packages.yml covers both
  unchanged.

Violations surfaced by the newly-working gate (~85) fixed across the repo:
- server.ts plan validation: string throws → Error (the 400 handler now
  returns err.message, wire format unchanged); async request handler and
  base-agent SIGINT/SIGTERM hooks routed through explicit `void`; worker
  onStart hooks made sync (interface widened to void | Promise<void>).
- Web: floating/misused promises void-wrapped, getTask casts typed as
  TaskStatus (enum-safe comparisons), JSON.parse through unknown,
  base-to-string guard in chunk-reload-guard, redundant union dropped.
- Rest is eslint --fix (unnecessary assertions, type-only imports).
- permit.ts: autofix removed a load-bearing 0x-assertion (lint project !=
  tsc project) — caught by typecheck, replaced with explicit annotations.

Gates: root lint 0, web lint 0, pnpm -r typecheck clean, demo-agents
211/211, adapter-evm 37/37, web build clean. Behavior-neutral — no
redeploy needed; changes ride along with the next Fly/Pages deploy.

+++ apps/demo-agents/src/orchestrator/guards.ts
@@ -62,7 +62,7 @@ export async function checkSponsorStatus(
     abi: ERC20_BALANCE_OF_ABI,
     functionName: 'balanceOf',
     args: [sponsor],
-  })) as bigint;
+  }));
 
   // Two bands: below min → critical (reject). Below 2×min → low (accept + warn).
   let level: SponsorStatus['level'] = 'healthy';
